decopule ChildClassMethodDecorator, and DependencyClassMethodDecorator
